+++ receiver/statsdreceiver/internal/parser/statsd_parser_test.go
@@ -2135,7 +2135,7 @@ func TestStatsDParser_IPOnlyAggregation(t *testing.T) {
 	value := metrics[0].Metrics.
 		ResourceMetrics().At(0).
 		ScopeMetrics().At(0).
-		Metrics().At(0).Sum().DataPoints().At(0).IntValue()
+		Metrics().At(0).Sum().DataPoints().At(0).DoubleValue()
 
-	assert.Equal(t, int64(4), value)
+	assert.Equal(t, 4.0, value)
 }
